Implementation and testing usecases of Invoice

+++ src/main/java/com/sofkau/dddchanllenge/domain/invoice/values/Prepayment.java
@@ -4,15 +4,15 @@ import co.com.sofka.domain.generic.ValueObject;
 
 import java.util.Objects;
 
-public class Prepayment implements ValueObject<Integer> {
-    private final Integer value;
+public class Prepayment implements ValueObject<Double> {
+    private final Double value;
 
-    public Prepayment(Integer value) {
+    public Prepayment(Double value) {
         this.value = value;
     }
 
     @Override
-    public Integer value() {
+    public Double value() {
         return value;
     }
